Move some interfaces to kopia package
Searching for new bases is mostly kopia-specific but is going to require
the store interface. With the way packages and interfaces are currently
laid out it makes it impossible to easily get the interface added to the
inject package without a cycle.
