+++ src/internal/operations/inject/inject.go
@@ -6,13 +6,11 @@ import (
 	"github.com/alcionai/corso/src/internal/common/idname"
 	"github.com/alcionai/corso/src/internal/common/prefixmatcher"
 	"github.com/alcionai/corso/src/internal/data"
-	"github.com/alcionai/corso/src/internal/kopia"
 	"github.com/alcionai/corso/src/pkg/account"
 	"github.com/alcionai/corso/src/pkg/backup/details"
 	"github.com/alcionai/corso/src/pkg/control"
 	"github.com/alcionai/corso/src/pkg/control/repository"
 	"github.com/alcionai/corso/src/pkg/fault"
-	"github.com/alcionai/corso/src/pkg/path"
 	"github.com/alcionai/corso/src/pkg/selectors"
 )
 
@@ -29,28 +27,6 @@ type (
 		) ([]data.BackupCollection, prefixmatcher.StringSetReader, error)
 
 		Wait() *data.CollectionStats
-	}
-
-	BackupConsumer interface {
-		ConsumeBackupCollections(
-			ctx context.Context,
-			bases []kopia.IncrementalBase,
-			cs []data.BackupCollection,
-			pmr prefixmatcher.StringSetReader,
-			tags map[string]string,
-			buildTreeWithBase bool,
-			errs *fault.Bus,
-		) (*kopia.BackupStats, *details.Builder, kopia.DetailsMergeInfoer, error)
-	}
-
-	RestoreProducer interface {
-		ProduceRestoreCollections(
-			ctx context.Context,
-			snapshotID string,
-			paths []path.RestorePaths,
-			bc kopia.ByteCounter,
-			errs *fault.Bus,
-		) ([]data.RestoreCollection, error)
 	}
 
 	RestoreConsumer interface {
